Moves RenderedProposal component to parent directory
